customer id long

+++ Aquasip/Controllers/CustomerManagesController.cs
@@ -69,7 +69,7 @@ namespace Aquasip.Controllers
         }
 
         // GET: Suppliers/Edit/5
-        public async Task<IActionResult> Edit(int? id)
+        public async Task<IActionResult> Edit(long? id)
         {
             if (id == null)
             {
